Move repo to be a component instead of a view.
